implementing review and tv rating

+++ database/migrations/2026_06_24_055616_create_tv_lists_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('acara');
             $table->string('thumb')->nullable();
             $table->enum('category', ['unsorted'])->default('unsorted')->nullable();
-            $table->bigInteger('rate')->nullable();
+            $table->float('rate')->nullable();
             $table->string('url');
             $table->enum('status', ['online', 'offline'])->default('online');
             $table->bigInteger('watched')->nullable()->default(0);
